app/requirements.txt
fastapi~=0.109.0
uvicorn~=0.23.1
requests==2.32.0
prometheus_fastapi_instrumentator~=6.1.0
python-dotenv
jinja2